Adding k8s yamls and renamed the package
