+++ k8s/app.yaml
@@ -1,0 +1,38 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: team-apps
+  namespace: dev
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: team-apps
+  template:
+    metadata:
+      labels:
+        app: team-apps
+    spec:
+      serviceAccountName: team-apps
+      restartPolicy: Always
+      containers:
+        - name: app
+          image: example/team-apps:1.0
+          env:
+            - name: SPRING_PROFILES_ACTIVE
+              value: cloud,async
+          ports:
+            - name: web
+              containerPort: 8080
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: team-apps-service
+  namespace: dev
+spec:
+  selector:
+    app: team-apps
+  ports:
+    - port: 8181
+      targetPort: web
