+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'blog-show-pods'
+rootProject.name = 'team-apps'
