+++ src/main/java/com/example/k8s/AsynchronousTeamAppsService.java
@@ -1,4 +1,4 @@
-package com.example.k8spods;
+package com.example.k8s;
 
 import io.kubernetes.client.informer.SharedIndexInformer;
 import io.kubernetes.client.informer.SharedInformerFactory;
