+++ src/main/java/com/example/k8s/TeamApp.java
@@ -1,4 +1,4 @@
-package com.example.k8spods;
+package com.example.k8s;
 
 import lombok.Builder;
 import lombok.Value;
