+++ src/main/java/com/example/k8s/TeamAppsController.java
@@ -1,4 +1,4 @@
-package com.example.k8spods;
+package com.example.k8s;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
